lcitool: move ansible/group_vars/ to facts/targets/
The ansible/group_vars/ directory is not directly related to what is
passed to Ansible in group variables; those also include packages.
Furthermore, it is also used to generate Dockerfiles which does not
involve Ansible.  Move it to another name which matches the name
of the variables in the code.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>