inventory: merge dictionaries from shared_facts into target facts
Reverse the direction in which facts[target] is created.  Instead
of starting with a copy of the shared facts and overriding them
with target-specific facts, add the missing shared facts to
facts[target].

With the old algorithm, subdictionaries could not be merged; the
change enables that change, for example to allow different paths
to come from different YAML files.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>