commandline: Add workload-dir optional argument
This is the absolute path to the directory which is
mounted as the "datadir/" folder in the user's
home directory inside the container.

The path is expected to contain all the files and
folders required to perform any workload in the
container. In order to build a project, the path
passed to --workload-dir would contain the source code
of the project and other necessary files and folder.

The logic behind providing a --workload-dir argument is to
provide a simpler interface to a directory which would
contain all the data required to perform an operation inside
the container rather than having to pass different arguments
such as the home directory, working directory, source-code
directory et.c to be mounted inside the container.

e.g --workload-dir /tmp/container_dir

Signed-off-by: Abdulwasiu Apalowo <abdulwasiuapalowo@gmail.com>