commandline: Add script optional argument
The script argument is the path of the executable
file which is used to run the workload contained in
--workload-dir CLI option. It can either be a
relative or absolute path.

It is the orchestrator of the content of WORKLOAD_DIR
(directory which contains all the files and folders
required to perform any workload on the container).

NB: The file is mounted into the container and can be
run in the container as a standalone workload (e.g an
"echo hello world" script)

e.g --script script.sh

Signed-off-by: Abdulwasiu Apalowo <abdulwasiuapalowo@gmail.com>