formatters: add a ShellBuildEnvFormatter class
The new formatter class generates a shell script that contains a
function 'install_buildenv'. When called this shell function will
install all the packages needed to build the project. The shell
script will also export any relevant environment variables.

In terms of the build dependancies, the result should be functionally
equivalent to what's present in the dockerfiles.

In theory the dockerfiles could actually just source these shell
scripts and call the functions, but then the dockerfiles would
no longer be self-contained. Thus we just maintain the commonality
in the formatter classes, and accept the duplication in the
generated outputs.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>