ansible: update freebsd-13 to latest minor release
After installing meson+python, symbol errors are present:

ld-elf.so.1: /lib/libc.so.7: version FBSD_1.7 required by /usr/local/lib/libpython3.9.so.1.0 not found

This is usually a sign that the port was built on a newer version
of the FreeBSD image. Since 13.1 is now available in Cirrus CI
we should update it to regardless.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>