manifest: allow overriding upstream project namespace as a variable
Some of the job rules and scripts refer to the upstream namespace of
the project. With the namespace hardcoded in this way, it can be
difficult to test CI changes prior to them being merged upstream.

This commit introduces a 'RUN_UPSTREAM_NAMESPACE' variable that can
be set when triggering a pipeline. That way CI maintainers can create
a fork of their fork, and treat the 1st level fork as "upstream" for
the purposes of testing changes in their 2nd level fork.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>