manifest: generate shell buildenv for all Linux targets
Currently all Linux targets build and use containers for the CI jobs. A
following commit is going to eliminate creation of cointainers in repo
forks, in favour of running a shell script for populating the buildenv
on the fly. This change causes the manifest command to generate suitable
shell scripts in ci/buildenv/*.sh

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>