manifest: stop generating containers in project fork repositories
With the changes in the free tier, users are limited to 5 GB of storage
quota. This is quickly consumed when CI pipelines publish 10-20
containers in the fork. The only option is to stop publishing containers
in fork repositories.

With this change, containers are only published in the upstream project
after code is committed to the default branch. There now needs to be two
sets of build jobs, one set will run using the pre-built container from
upstream. The other set will use the shell buildenv script to
dynamically populate a fresh build environment from the 3rd party distro
base container image.

Build jobs running upstream post commit on the default branch can use
the pre-built containers, but those running other (stable) branches
must always populate a fresh environment. In forks, pipelines will
always populate a fresh environment, as it can't be determined through
CI env variables alone, whether there are CI changes vs upstream.
The 'changes:' clause is not usable for pushes in forks, as it
reports on the delta vs the previous fork branch state, not the
delta vs upstream.

As well as saving storage quota in forks there are some other benefits.
Pipelines testing commits which do not modify CI config will now be
faster, since we're always skipping the container jobs. This new scheme
will also unlock the ability to trigger pipelines in merge request
context. Previously this was impossible, as the container jobs would
publish containers for code that is yet to be reviewed or merged.
As mentioned earlier, it is also now possible to run CI for upstream
stable branches, since it won't mistakenly use containers associated
with latest mainline.

The effect of this change is a variety of different jobs will
run depending on scenario

  - Push upstream default branch with CI changes

     * $TARGET-container
     * $TARGET-prebuilt-env

  - Push upstream default branch without CI changes

     * $TARGET-prebuilt-env

  - Push upstream non-default branch

     * $TARGET-local-env

  - Push fork

     * $TARGET-local-env

NB, that the decision to use a prebuilt vs local env build jobs
is per target, when triggered based on CI changes. IOW if a fork
dockerfile for fedora-rawhide-mingw64 changes, only that target
will use local env, remaining targets will continue with the
prebuilt env.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>