manifest: switch to triggering pipelines from merge requests
Currently pipelines are triggered from branch pushes, even though they
are set as a pre-condition for merge request acceptance. This is
wasteful of contributors' limited CI quota as often there is a need to
push separately from submitting code for merge. To stop wasting CI quota
this switches to only trigger pipelines when a merge request is opened.

This also enables projects to make use of merge trains, whereupon the
CI is re-run after the merge request is accepted, to prove it still
passes even if HEAD has moved on.

Contributors can force a pipeline on push by setting  'RUN_PIPELINE=1'
as an environment variable.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>