containers: Implement "_build_args" method
Generate some of the options needed to run workloads
in a container. It replaces some of the logic in
`libvirt.git/ci/Makefile` and adds a simpler interface
with the use of "datadir".

If `datadir` is supplied, it is mounted to the
"datadir/" directory in the user's home directory inside
the container.

if `script` is supplied, it is copied into a temp. directory,
and also mounted to the "script" file inside the
container. It is required to kick-start the operation.

NB: The working directory is changed to the home directory if
--workload-dir or --script is supplied to the CLI options

/etc/{passwd,group} files are copied to a temporary dir
(temp_dir). The respective files are then mounted to
/etc/{passwd,group} in the container.

Signed-off-by: Abdulwasiu Apalowo <abdulwasiuapalowo@gmail.com>