lcitool: install: Introduce a new installer module
This module will host the virt-install code which currently lives in
the application module and more. The reason for creating a new module
(and a package as well) is to allow future extensibility of the
VirtInstall installer class, so let's start by introducing a very basic
skeleton which future patches can build upon.

Signed-off-by: Erik Skultety <eskultet@redhat.com>