configs: Move the cloud config to the install package
So far, the config only served as an example of a config users might
get inspired by when initializing their own cloud-init powered images
(originally created by lcitool). Ultimately, we'd like to utilize
cloud-init for VM installs ourselves, so it makes sense to move the
config file to the new 'install' package from example. Note however,
that actually turning this config into something useful is
beyond the scope of this patch, but moving it now allows us to
decrease the directory nesting of the lcitool/configs directory
immediately.

Signed-off-by: Erik Skultety <eskultet@redhat.com>