facts/projects: fix libvirt Go project names to match repo
Previously the manifest had relied on the $CI_PROJECT_NAME variable to
build a path to the upstream container registry. This was incorrect
though, because the repo fork may have had its name changed relative
to upstream.

This the current manifest code uses the lcitool project name to identify
the upstream cohntainer registry. This means that project names need to
match the git repo names, but we never updated the libvirt Go projects
after we switched development to new repos to enable Go modules usage.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>