gitlab: allow jobs to be interrupted
If a job has 'interruptible: true', then when a new pipelines starts on
a branch, it can auto-cancel any previous pipelines that are still
running. This will save CI quota for merge request pipelines when users
re-push the branch for the MR multiple times in a short period, by
cancelling pipelines for outdated branches.

The only jobs we don't allow to be cancelled are the container build
jobs. These only run on upstream push / scheduled / manual jobs, and
should always be allowed to run to completion, otherwise we risk an
inconsistent set of containers upstream. This is OK, as it is unlikely
we'll be in a situation where concurrent pipelines would be running
upstream.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>