install: Make the cloud-init.conf a template and update it
Firstly, make it clear that the config file is a dynamically edited
template by using the '.in' suffix as well as a top level commentary.
Secondly, update the contents of the template a bit so that it meets
our future needs for the very initial configuration of a machine which
we'll then update with Ansible.

Notable changes:
- set an initial plaintext root password
- allow initial SSH password authentication

Signed-off-by: Erik Skultety <eskultet@redhat.com>