install: Add a class method to instantiate URL-based installer object
Add a convenience class method allowing us to quickly instantiate a VirtInstall
object tailored to URL-based installs - all that is needed is to
execute 'run()' and a guest will be installed from a URL.

It is still possible to instantiate and configure a VirtInstall
instance 100% manually, but having a shortcut providing all the correct
virt-install arguments is convenient. By having a class method we'll be
able to add more of them to support various install scenarios.

Signed-off-by: Erik Skultety <eskultet@redhat.com>