install: Add a class method to instantiate an image-based installer object
Some of the args we need virt-install to take differ from the ones we
pass onto it when doing URL-based installs. In previous patches we
added a shortcut class method returning a configured VirtInstall
instance ready to execute the URL-based install. Do the same for
image-based install. Given that we have to do more prep steps for
cloud-init as well as make sure we downloaded the base image
successfully, the code is much more complex than its URL counterpart,
e.g. we're making sure that the downloaded image is put properly into
the given libvirt storage pool.

Signed-off-by: Erik Skultety <eskultet@redhat.com>