install: Introduce a cloud-init handling module
For now, we only need to deal with cloud configs, since virt-install
is going to take care of the rest like creating and mounting a
NoCloud ISO source to the guest for us.

The CloudConfig class is just a glorified dictionary with some
convenience methods. The reason why UserDict was selected for the base
class instead of just plain dict is that inheriting from dict is fine
when only adding methods/attributes, not when any of the existing or
magic methods needs to be changed - dict just wasn't meant to be
tampered with and for that purpose UserDict was created in the
collections module, which is fully integrated and meant to be
subclassed. To illustrate the problem, if one tries to YAML format a
custom dict-subclassed object, this appears in the result:

!!python/object/new:lcitool.install.cloud_init.CloudConfig
dictitems:
  <dict_key>: <dict_value>
  ...

Apparently YAML checks the object type which is surely not going to be
<class dict>. Since UserDict is fully integrated, the problem above
doesn't exist and everything works as expected. The price to pay for
UserDict is that it's pure Python and so significant performance drop
is expected, but since we don't expect a ton of lookups on the
CloudConfig object, we're totally fine.

Signed-off-by: Erik Skultety <eskultet@redhat.com>