util: Add abstraction to represent SSH keys
We'll need to figure out what the private key is when trying the SSH
connection with cloud-init based installs, so let's add a keypair
abstraction. Note that this abstraction is also handy if all you need
is a pubkey, because everything is stored in a single place and given
that it's very likely one will need to read the public key, this
abstraction has a helper to quickly return the public key as string.

Signed-off-by: Erik Skultety <eskultet@redhat.com>