libvirt_wrapper: Introduce an abstract LibvirtObject wrapper base class
We'll need to work with more libvirt objects than just the existing
convenience LibvirtWrapper itself and so we need to describe and wrap
those individual libvirt object types with an abstract class.

This abstract class mandates that the underlying libvirt object is
always accessible and available to be used via the 'raw' attribute.

Signed-off-by: Erik Skultety <eskultet@redhat.com>