install: Define an SSH wait callback
So, ideally we'd like to use the industry standard - cloud-init - for
this, however since libvirt creates its own firewalld zone for VMs we'd
have to consider the problem of intra-zone traffic and rules. In
practice, this means that we'd have to make sure that the default zone
on the host has the whole port range accepted since the on-demand HTTP
cloud-init server would use a random port every time. That is something
that is beyond the scope of lcitool and given lcitool tries to make
creation and setup of a throwaway VM as simple as possible, expecting
users to setup their firewall in a certain way we expect is simply not
viable. Therefore, this patch replaces the cloud-init phone_home
approach with a plain SSH-poll in a loop type of solution. Although not
bullet-proof, it is a simple solution that works everywhere out of the
box with the only significant problem being the static timeout after
which we expect the machine to be fully booted which may not be
conservative enough (30s).

There's one important detail about exception handling with the Paramiko
SSH library - we need to catch 2 types of exceptions (OSError,
paramiko.NoValidConnections) depending on how exactly paramiko handles
SSH socket errors when we're trying to connect to a machine that is not
fully booted yet - this is a problem of paramiko and we cannot do
anything about it.

Signed-off-by: Erik Skultety <eskultet@redhat.com>