install: image: Add a method to download the image from internet
With the data from libosinfo, we just need to take the image URL and
pull it from the remote source and save it in cache with temporary
suffix. The reason for that is there might already be a VM running from
an older base cloud-init image we're essentially trying to replace.
To avoid the actual act of replacement we'll use a random string for
the destination file. Our metadata about the base image in question is
dumped along with the image itself.

Signed-off-by: Erik Skultety <eskultet@redhat.com>