install: Introduce a new InstallationNotSupported exception
We have a few occurrences where we error out with
    "Host XYZ doesn't support installation."

Firstly, the fact a host cannot be installed is tied to the underlying
target and not the host itself. Secondly, we can do better and actually
define a corresponding exception type now that we separated the
installation stuff to a new module. This patch fixes both issues
mentioned above.

Signed-off-by: Erik Skultety <eskultet@redhat.com>