formatter: add inventory field
Prepare for making the Inventory not a singleton: add an inventory
to each formatter and push the creation of the inventory up to
the same code that creates the formatters.

For now all formatters point to the same object, but this can be changed
as soon as each user of formatters receives the Inventory object from
its creator.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>