manifest: add inventory field
Prepare for making the Inventory not a singleton: add an inventory
to the Manifest object and push the creation of the inventory up
the same code that creates the manifest.

For now all manifests point to the same object, but this can be changed
as soon as each user of manifests receives the Inventory object from
its creator.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>