tests: define fixtures corresponding to singletons
Instead of using global variables or singletons, make the tests use
pytest fixtures.  Module-level caching to avoid excessive test
runtimes once singletons are removed, but the fixture are not defined
in conftest.py so that this can be changed in specific test modules
if desired.

Sometimes, the objects are needed to list the test parameters, and
therefore must be instantiated at collection time.  In that case, the
globals are named _NAME and the fixtures simply return them; however,
for consistency and cleanliness the tests never ever access the globals
themselves.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>