commandline: Add engine optional argument
This patch adds an optional argument called
"engine". The default value is "podman".

It will be used by the "run" and "shell"
sub-commands to determine the engine to use(from the
list of available engines) for image build and/or
running workloads on containers.

e.g --engine podman OR --engine docker

Signed-off-by: Abdulwasiu Apalowo <abdulwasiuapalowo@gmail.com>