facts: introduce macos 13 Cirrus CI target on aarch64
Cirrus CI has replaced their infrastructure for runing macOS images and
now targets the M1 hardware. The original x86_64 hardware is no longer
supported and users are required to stop using it prior to Jan 1st 2023.
The Ventura release is only supported for M1 hardware in Cirrus CI and
is the new default

https://cirrus-ci.org/blog/2022/11/08/sunsetting-intel-macos-instances/

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>