inventory: add projects field
Make the inventory represent the global state loaded from the facts/
directory and the data directory.  All places that need an inventory
*and* projects object already have the inventory, so it's easy to
retrieve the projects object from there.

For now all inventories point to the same object, but this can be changed
as soon as each user of Inventory receives the Projects object from
its creator.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>