projects: do not pass target facts to PackageFactory constructor
The target information is passed to PackageFactory half at
construction time and half at get_package time.  Keep only
the latter; the facts are only used to cache the mapping keys,
but really it's just as simple to build the keys when needed.
Alternatively, they could also be cached in the target object.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>