install: image: Load base image metadata and instantiate pool of images
This patch scans local cache and loads base image metadata telling us
how many and what kind of base images are available for provisioning
without the need to download the same kind of target OS image
repeatedly. If more images (and metadata) for the same target OS exist,
only the most recent is considered, the rest is skipped - the old
images can still be in active use though so we're not touching them.

Signed-off-by: Erik Skultety <eskultet@redhat.com>