targets: alias lcitool.inventory.Inventory to lcitool.targets.Targets
As a first step in separating Targets from Inventory, make Targets
just an alias of Inventory.  All clients that do not care about the
host facts are changed to use "targets" and "Targets" respectively
as the variable and the class name.

This makes little sense code-wise, but it makes it possible to
separate the mechanical part of the change.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>