rename Target to BuildTarget
With the introduction of a Targets class, using "Target" for the OS+arch
pair becomes confusing.  In the future "Target" could be used possibly
for the object that Targets holds onto, similar to Project vs. Projects;
right now it does not make much sense, because Target would be a
glorified dict holding the facts with no other projects or methods.
Still, avoid the confusion *now* by renaming the class to BuildTarget.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>