reject docutils>=0.18 from pypi
https://github.com/readthedocs/sphinx_rtd_theme/blob/master/setup.cfg shows that
sphinx-rtd-theme does not want versions of docutils starting with 0.18.  For
distro-installed packages we can assume that the packager knew this, but
when installing from PyPI the package version needs to be constrained by
hand.

In order to do this, add the docutils package explicitly to the qemu project file.
If a target needs docutils from PyPI (which is never the case currently, but
it happened to me while playing with CentOS 8's Python 3.8 packages), the
output of lcitool will force the installation of docutils<0.18 instead of
picking whatever release it wants and then failing:

   Step 3/9 : RUN /usr/bin/pip3.8 install                     meson==0.63.2                     pillow                     sphinx                     sphinx-rtd-theme
   ...
   Collecting docutils<0.20,>=0.14
     Downloading https://files.pythonhosted.org/packages/93/69/e391bd51bc08ed9141ecd899a0ddb61ab6465309f1eb470905c0c8868081/docutils-0.19-py3-none-any.whl (570kB)
   ...
   ERROR: sphinx-rtd-theme 1.1.1 has requirement docutils<0.18, but you'll have docutils 0.19 which is incompatible.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>