lcitool container: Add "build" subcommand
This adds the "build" subcommand which is used to
build container image from args.projects and
args.target.

The Dockerfile string generated from using the
`DockerfileFormatter` class is copied to a temporary file
which is passed to the build command.

Tags are generated by prefixing the target name with
"lcitool.", e.g "lcitool.fedora-36" to prevent trashing
the output of "podman images" with incomprehensible tags.

If an image is present, it is removed before starting another
build to ensure the image is built from scratch.

Ensure that (--target & --projects) arguments are passed with
"build" subcommand.

Signed-off-by: Abdulwasiu Apalowo <abdulwasiuapalowo@gmail.com>