configs: kickstart: Erase all existing partitions
In the comment we say that we erase everything except that we use
'--none' with the 'clearpart' directive. It just so happens that this
has always been a NOP since the default for VMs created by lcitool
is to always create a new qcow2 disk. FWIW, use the right option so
that the statement is actually correct.

Signed-off-by: Erik Skultety <eskultet@redhat.com>