projects: use a different cache key for native vs. cross targets
Right now, the same key is used for both a native and a cross x86_64 target.
The two however can have different packages when the build is done on a non-x86_64
host.

To fix this, change the cache key for native targets to not include the "-x86_64"
suffix.  The set of packages for native targets, as far as lcitool is concerned,
is the same for all host architectures, so there is no reason to mention x86_64.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>