facts: workaround bogus retirement of capstone from EPEL9
EPEL9 retired capstone as soon as it was added to c9s for
the future RHEL-9.2 release. This dooms anyone still using
RHEL < 9.2 who still needs capstone from EPEL, so we
workaround by manually picking the most recent version

This relies on the fact that DNF accepts URLs to package
downloads, as an alternative to plain package names. This
is good enough for generating containers, but could break
if we get other code in lcitool that expects it to always
be a literal package name.

https://bugzilla.redhat.com/show_bug.cgi?id=2124181
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>