facts: adapt to Alpine change from php8 to php81
In Alpine 3.16, there are php81 packages in parallel with php8

In Alpine Edge, there are php82 packages in parallel with php81,
and no php8.

We can't use the php82 packages yet, however, since the pecl-imagick
package isn't available.

So this switches to the php81 stack for 3.16 and Edge. Edge will
need updating again to php82 when it is fully available.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>