containers: publish to 'latest' tag instead of 'master'
Historically we use the CI_COMMIT_REF_SLUG for the docker tag, which is
the branch name. This meant that when testing container builds from
non-default tags/branches we didn't pollute the primary published
content.

When we switched to use merge requests for triggering CI in:

  commit 7a7402d259db6e75b2a6fbb1523b42a2d18fabfa
  Author: Daniel P. Berrangé <berrange@redhat.com>
  Date:   Thu Feb 10 17:43:56 2022 +0000

    gitlab: switch to trigger jobs against the merge request by default

we stopped publishing containers for anything except a push to the
default branch. As such there's no reason to use CI_COMMIT_REF_SLUG,
we can just use a fixed tag name.

The docker default is to use ':latest', so adopt that instead of
':master'.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>