lcitool: fix compat with podman 4.4.0 and docker
The '{{ json . }}' syntax has returned a single JSON
document with an array of entries for images historically
with podman.

Since podman 4.4.0 this now returns a concatenation of
JSON documents, one document per image. Our code is not
able to parse a concatenation of documents.

This was done to fix docker compatibility[1], because

  {{ json .}}

was supposed to be equivalent to

  {{ . | json }}

while podman mistakenly made it equivalent to

  {{ json }}

IOW, we were already broken in terms of docker compat
too.

Since we want the single document, we need to use plain
'json' or '{{ json }}'. This picks the former since it
does not appear we have a need to use any advanced
features of the Go template syntax.

[1] https://github.com/containers/podman/issues/16436
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>