Add pyproject.toml
Replace bare setup.py. The required version of setuptools set by this
patch is not a random version number - it is the version where
setuptools' pyproject.toml parsing gained support for recursive
globbing of package data without which we would not be able to
enumerate package data as simply as "<package>/datadir/**".

Signed-off-by: Erik Skultety <eskultet@redhat.com>