commandline: Set program name explicitly
The flat project hierarchy allows python to run the project directly
with 'python3 -m' and Python will automatically import the package.
However, that way it'll run the project via the __main__.py module
which in turn will set argv[0] to __main__.
Setting the program name explicitly in argparse is useful nonetheless.

Signed-off-by: Erik Skultety <eskultet@redhat.com>