lcitool: Introduce a __main__ module
The thing with transitioning from setup.py to pyproject.toml build
strategy is that we now must specify a list of all entry points for the
project, IOW whatever is going to become the CLI executable. With
setup.py we got away by having a bin/lcitool script which would get
installed directly, but that's not how setuptools builds work after PEP
517 [1]. The build backend now creates the executables automatically as
wrappers over the entry points that are required to be listed in the
project's build config. There's one more requirement - the entry point
needs to be importable through Python's standard means. Since
bin/lcitool is not part of the lcitool package, hence not easily
importable as a module, we need to have its logic elsewhere. One of the
standard ways of doing that is by creating a __main__.py module which
python also automatically looks for when instructed to run the project
directly from its repo.

[1] https://peps.python.org/pep-0517/

Signed-off-by: Erik Skultety <eskultet@redhat.com>