util: Add a 3.8 compatibility helper over importlib.resources.files
importlib.resources.files was introduced in Python 3.9. However, we
support Python 3.8+ which means that we need to add a compatibility
helper to query package resources properly. Normally we'd just go with
importlib.resources.path, but that API is lacks any kind of
practicality as it only accepts filenames as a resource, rejecting
relative paths, strings with directory delimiters or even directories.
So, if we want to iterate over a bunch of data files we ship along with
lcitool, we have to use other means to do that - get the imported
package directory and join that path with whatever we're looking for.
This solution doesn't work with resources not physically on the file
system, IOW with resources shipped as a ZIP which isn't our case
anyway.
Note that this patch can be reverted once we move onto 3.9+ in the
future.

Signed-off-by: Erik Skultety <eskultet@redhat.com>