tests: Introduce class and module scoped equivalents for monkeypatch
If one tries to combine a class-scoped or a module-scoped fixture that
happens to depend on monkeypatch, this is what they'll get:

    "You tried to access the function scoped fixture monkeypatch with
    a module scoped request object"

Sometimes one wants to monkeypatch to carry over the override to
further test functions and that's where fixture scopes come in. To
remediate the issue, introduce our own global fixture helpers with
conforming to different scopes.
At the same time, apply the improvement to test_containers.py.

Signed-off-by: Erik Skultety <eskultet@redhat.com>