install: Store the scheme for unattended install in target's config
When installing from a URL, we relied on the hardcoded combinations of
unattended install schemes and their corresponding targets.
The problem with this approach is that even though we now support
adding new targets or target overrides using the data dir one cannot
actually install a VM instance of that target unless the OS name is
something we already know about in upstream. It may sound like adding
the new target to the upstream codebase would be the way to go instead,
but if you consider the target being RHEL which, at least from
unattended install's perspective, cannot be simply added to the list of
supported targets, we need to look for a different mechanism to support
downstream use cases involving lcitool and these enterprise distros,
provided the user has access to repositories and os install trees.

This patch solves the problem by introducing a new attribute
'install.unattended_scheme' in the target's YAML profile. This will
then determine the name of the corresponding install config file to be
injected.

Signed-off-by: Erik Skultety <eskultet@redhat.com>