util: Use pathlib native way of resolving user's home
Instead of doing os.environ["HOME"] use pathlib's native way of
resolving the home directory: Path.home()
The reason for this is that in test environments "HOME" might not be
set which results in a vague "KeyError: 'HOME'" error.
If pathlib fails to resolve home, it raises
    'RuntimeError: Can't determine home directory'. The benefit is that
instead of looking at $HOME pathlib tries to use os.expanduser() to
construct the home path which is more robust and the error is also more
verbose than a plain "KeyError".

Signed-off-by: Erik Skultety <eskultet@redhat.com>