tox: Allow running with custom pytest options with {posargs}
Tox allows passing custom command arguments with {posargs}. This is
useful to us because we have our own custom pytest option
'--regenerate-output' which we need to be able to pass to pytest.
The way this works is you run tox as:
    $ tox -- <--custom_option>

and tox will substiture <--custom_option> everywhere where {posargs} is
mentioned. That would be a problem if we ever allow running multiple
tox environments with custom options, but in that case limiting the
environments to be run like the following would again solve the issue:

    $ tox -e lint -- --color never
    $ tox -e py38 -- --regenerate-output

Signed-off-by: Erik Skultety <eskultet@redhat.com>