gitlab: pass CI_MERGE_REQUEST_REF_PATH to Cirrus jobs
The Cirrus CI jobs are passed CI_REPOSITORY_URL and CI_COMMIT_REF_NAME
variables. The former is the repository against which the job is being
run and the latter is the branch name of the code to test. The latter,
however, is only valid when testing already merged code either in
upstream or in a fork.

If testing a merge request then CI_REPOSITORY_URL refers to the upstream
repo, while CI_COMMIT_REF_NAME refers to the branch in the contributor's
fork. Cirrus CI is of course unable to fetch CI_COMMIT_REF_NAME from
upstream since it is the wrong repository.

Fortunately the commits assocaited with any merge request are also
copied into upstream and made available at merge-requests/$NUM/head.
This ref path is present in the CI_MERGE_REQUEST_REF_PATH env variable,
and so this should be exposed to Cirrus CI too.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>