tests: utils: Introduce a class representing a diff operand
The idea is to be able to compare all datatypes we produce in a
generic manner interchangeably. In other words, we need to be able to
compare lists, dicts, strings, files or even strings with files
transparently to the caller. How are we going to do that? Since our
native serialization format is YAML, we'll serialize everything apart
from raw strings to YAML and compare those the same way as if two files
were compared using diff. To implement the idea, difflib, which is part
of the standard library, is used.

The interface to compare the actual and expected outputs was
borrowed from Python sets and so follows the "A.diff(B)" mantra where
A is the actual produced output and B is the expected one. Both need to
be of type DiffOperand (introduced by this patch) which takes care of
any conversion needed. Unlike sets though, we can't use this the other
way around because then we'd get difflib's uninified_diff confused on
which is which. Not that it wouldn't work, it's just the final diff
might be confusing. Sadly, there's no way to enforce this usage though.

The nice thing about this patch is that we'll now get something like
this if the outputs don't match:

  --- actual
  +++ expected
  @@ -48,7 +48,6 @@
   - firewalld-filesystem
   - flake8
   - flex
  -- foo
   - fuse
   - fuse3
   - fusermount
  @@ -317,7 +316,6 @@
   - xsltproc
   - xz
   - yajl
  -- zfoo
   - zip
   - zlib
   - zlib-static

Signed-off-by: Erik Skultety <eskultet@redhat.com>