tests: utils: Introduce a module-private method _assert_equal
So, this function will eventually replace the existing assert_equal
function. However, unlike the original, this one is not to be used
directly and instead will be wrapped by a global fixture introduced in
a future patch. The reason for that is simple. To give the developer
the most convenient output possible, we're going to dump both the
actual and expected outputs to the disk in order to aid them with
debugging problems. Where do we dump the files though? The natural
answer is "wherever pytest would create them".
Indeed pytest has the 'tmp_path_factory' fixture which can create
directories on demand, but what is the name for such a directory?
Pytest has another fixture for that - request. And now every single
test has to depend on 2 fixtures to get the test's name and to create a
temporary directory using that name to dump data in there on failure.
A test writer should not be concerned by this and they should only ever
care about comparing the actual and expected outputs, that's it.

Signed-off-by: Erik Skultety <eskultet@redhat.com>