tests: conftest: Introduce a global fixture assert_equal
This fixture is going to be used by every single positive tests.
Because it's global, it's going to be available everywhere, and because
it return a higher order function, it can only accept arguments for the
actual and expected outputs to be compared, everything else it
transparent to the test writer.

Signed-off-by: Erik Skultety <eskultet@redhat.com>