gitlab: Introduce envid
Short for "environment identifier", it's a combination of the
target name and, if applicable, the cross-build architecture.

Right now we're only using it to build the image name, so all
we're gaining is the removal of a tiny amount of duplication.
It will become much more important in a second.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>