gitlab: Use envid to look up resources
Using $NAME is fine when performing a native build, but when
cross-building we need to use $NAME-cross-$CROSS instead, or
we will pick up the wrong file and things will not work.

This currently affects all $CROSS-$NAME-local-env jobs, which
end up performing a native build instead of the expected
cross-build.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>